@@ -11,14 +11,15 @@
     "require": {
         "alex-kalanis/kw_paths": ">=3.0 <4",
         "ext-mbstring": "*",
-        "php": ">=7.2.0"
+        "php": ">=7.4.0"
     },
     "require-dev": {
-        "alex-kalanis/kw_storage": ">=4.0 <5",
+        "alex-kalanis/kw_storage": ">=5.0 <6",
+        "shipmonk/composer-dependency-analyser": "^1.4",
         "phpstan/phpstan": "^1.0",
         "phpstan/phpstan-phpunit": "^1.0",
         "friendsofphp/php-cs-fixer": "^3.0",
-        "phpunit/phpunit": ">=7.2 <=9"
+        "phpunit/phpunit": ">=8 <=9"
     },
     "suggest": {
         "alex-kalanis/kw_storage": "Local storage for accessing resources"
