@@ -20,13 +20,12 @@ use kalanis\kw_storage\StorageException;
  */
 class Basic extends ADirs
 {
-    /** @var IStorage */
-    protected $storage = null;
+    protected IStorage $storage;
 
     public function __construct(IStorage $storage, ?IFLTranslations $lang = null)
     {
         $this->storage = $storage;
-        $this->setLang($lang);
+        $this->setFlLang($lang);
     }
 
     public function createDir(array $entry, bool $deep = false): bool
@@ -58,7 +57,7 @@ class Basic extends ADirs
             }
             return $this->storage->write($entryPath, IProcessNodes::STORAGE_NODE_KEY);
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotCreateDir($entryPath), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotCreateDir($entryPath), $ex->getCode(), $ex);
         }
     }
 
@@ -68,7 +67,7 @@ class Basic extends ADirs
         $entryPath = empty($entryPath) ? '' : $this->getStorageSeparator() . $entryPath;
         try {
             if (!$this->isNode($entryPath)) {
-                throw new FilesException($this->getLang()->flCannotReadDir($entryPath));
+                throw new FilesException($this->getFlLang()->flCannotReadDir($entryPath));
             }
             /** @var array<string, Node> */
             $files = [];
@@ -109,7 +108,7 @@ class Basic extends ADirs
             }
             return $files;
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotReadDir($entryPath), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotReadDir($entryPath), $ex->getCode(), $ex);
         }
     }
 
@@ -153,7 +152,7 @@ class Basic extends ADirs
             }
             return true;
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotCopyDir($src, $dst), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotCopyDir($src, $dst), $ex->getCode(), $ex);
         }
     }
 
@@ -191,7 +190,7 @@ class Basic extends ADirs
             }
             return $this->storage->remove($src);
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotMoveDir($src, $dst), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotMoveDir($src, $dst), $ex->getCode(), $ex);
         }
     }
 
@@ -217,7 +216,7 @@ class Basic extends ADirs
             }
             return $this->storage->remove($path);
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotRemoveDir($path), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotRemoveDir($path), $ex->getCode(), $ex);
         }
     }
 
@@ -233,32 +232,12 @@ class Basic extends ADirs
 
     /**
      * @param string $file
-     * @throws FilesException
      * @throws StorageException
      * @return int
      */
     protected function getSize(string $file): int
     {
-        $content = $this->storage->read($file);
-        if (is_resource($content)) {
-            // a bit workaround
-            $tempStream = fopen("php://temp", "w+b");
-            if (false === $tempStream) {
-                // @codeCoverageIgnoreStart
-                throw new FilesException($this->getLang()->flCannotLoadFile($file));
-            }
-            // @codeCoverageIgnoreEnd
-            rewind($content);
-            $size = stream_copy_to_stream($content, $tempStream, -1, 0);
-            if (false === $size) {
-                // @codeCoverageIgnoreStart
-                throw new FilesException($this->getLang()->flCannotGetSize($file));
-            }
-            // @codeCoverageIgnoreEnd
-            return intval($size);
-        } else {
-            return mb_strlen(strval($content));
-        }
+        return strlen($this->storage->read($file));
     }
 
     /**
@@ -267,6 +246,6 @@ class Basic extends ADirs
      */
     protected function noDirectoryDelimiterSet(): string
     {
-        return $this->getLang()->flNoDirectoryDelimiterSet();
+        return $this->getFlLang()->flNoDirectoryDelimiterSet();
     }
 }
