@@ -4,9 +4,7 @@ namespace kalanis\kw_files\Processing\Storage\Files;
 
 
 use kalanis\kw_files\FilesException;
-use kalanis\kw_files\Interfaces\IFLTranslations;
 use kalanis\kw_paths\ArrayPath;
-use kalanis\kw_storage\Interfaces\IStorage;
 use kalanis\kw_storage\StorageException;
 
 
@@ -17,12 +15,6 @@ use kalanis\kw_storage\StorageException;
  */
 class Basic extends AFiles
 {
-    public function __construct(IStorage $storage, ?IFLTranslations $lang = null)
-    {
-        $this->storage = $storage;
-        $this->setLang($lang);
-    }
-
     public function copyFile(array $source, array $dest): bool
     {
         $src = $this->getStorageSeparator() . $this->filledName($this->compactName($source, $this->getStorageSeparator()));
@@ -40,7 +32,7 @@ class Basic extends AFiles
 
             return $this->storage->write($dst, $this->storage->read($src));
         } catch (StorageException $ex) {
-            throw new FilesException($this->getLang()->flCannotCopyFile($src, $dst), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotCopyFile($src, $dst), $ex->getCode(), $ex);
         }
     }
 
@@ -57,6 +49,6 @@ class Basic extends AFiles
      */
     protected function noDirectoryDelimiterSet(): string
     {
-        return $this->getLang()->flNoDirectoryDelimiterSet();
+        return $this->getFlLang()->flNoDirectoryDelimiterSet();
     }
 }
