@@ -4,12 +4,9 @@ namespace kalanis\kw_files\Processing\Volume;
 
 
 use kalanis\kw_files\FilesException;
-use kalanis\kw_files\Interfaces\IFLTranslations;
-use kalanis\kw_files\Interfaces\IProcessFiles;
+use kalanis\kw_files\Interfaces;
 use kalanis\kw_files\Processing\TPath;
 use kalanis\kw_files\Traits\TLang;
-use kalanis\kw_files\Traits\TStreamToPos;
-use kalanis\kw_files\Traits\TToStream;
 use kalanis\kw_paths\Extras\TPathTransform;
 use kalanis\kw_paths\PathsException;
 use Throwable;
@@ -20,21 +17,19 @@ use Throwable;
  * @package kalanis\kw_files\Processing\Volume
  * Process files in many ways
  */
-class ProcessFile implements IProcessFiles
+class ProcessFile implements Interfaces\IProcessFiles, Interfaces\IProcessFileStreams
 {
     use TLang;
     use TPath;
     use TPathTransform;
-    use TStreamToPos;
-    use TToStream;
 
-    public function __construct(string $path = '', ?IFLTranslations $lang = null)
+    public function __construct(string $path = '', ?Interfaces\IFLTranslations $lang = null)
     {
         $this->setPath($path);
-        $this->setLang($lang);
+        $this->setFlLang($lang);
     }
 
-    public function readFile(array $entry, ?int $offset = null, ?int $length = null)
+    public function readFile(array $entry, ?int $offset = null, ?int $length = null): string
     {
         $path = $this->fullPath($entry);
         try {
@@ -46,9 +41,9 @@ class ProcessFile implements IProcessFiles
                 $content = @file_get_contents($path);
             }
             if (false !== $content) {
-                return $content;
+                return strval($content);
             }
-            throw new FilesException($this->getLang()->flCannotLoadFile($path));
+            throw new FilesException($this->getFlLang()->flCannotLoadFile($path));
         } catch (Throwable $ex) {
             // @codeCoverageIgnoreStart
             throw new FilesException($ex->getMessage(), $ex->getCode(), $ex);
@@ -56,69 +51,99 @@ class ProcessFile implements IProcessFiles
         // @codeCoverageIgnoreEnd
     }
 
-    public function saveFile(array $entry, $content, ?int $offset = null): bool
+    public function readFileStream(array $entry)
     {
         $path = $this->fullPath($entry);
         try {
-            if (is_null($offset)) {  // rewrite all
-                $this->writeStream($path, $this->toStream($path, $content));
-            } else { // append from position
-                if (file_exists($path)) {
-                    $handler = @fopen($path, 'rb+');
-                    if (false === $handler) {
-                        // @codeCoverageIgnoreStart
-                        throw new FilesException($this->getLang()->flCannotOpenFile($path));
-                    }
-                    // @codeCoverageIgnoreEnd
-                    // must be extra - need that original file handler
-                    $result = $this->addStreamToPosition(
-                        $handler, // original content
-                        $this->toStream(
-                            $path,
-                            $content
-                        ),
-                        $offset
-                    );
-                    /** @scrutinizer ignore-unhandled */@fclose($handler);
-                    $this->writeStream($path, $result);
-                } else {
-                    $handler = @fopen('php://memory', 'rb+');
-                    if (false === $handler) {
-                        // @codeCoverageIgnoreStart
-                        throw new FilesException($this->getLang()->flCannotOpenFile($path));
-                    }
-                    // @codeCoverageIgnoreEnd
-                    $this->writeStream(
-                        $path,
-                        $this->addStreamToPosition(
-                            $handler, // no original content
-                            $this->toStream(
-                                $path,
-                                $content
-                            ),
-                            $offset
-                        )
-                    );
-                }
+            $handle = @fopen($path, 'rb');
+            if (false !== $handle) {
+                return $handle;
             }
-            return true;
+            throw new FilesException($this->getFlLang()->flCannotLoadFile($path));
         } catch (Throwable $ex) {
             // @codeCoverageIgnoreStart
-            throw new FilesException($this->getLang()->flCannotSaveFile($path), $ex->getCode(), $ex);
+            throw new FilesException($ex->getMessage(), $ex->getCode(), $ex);
+        }
+        // @codeCoverageIgnoreEnd
+    }
+
+    public function saveFile(array $entry, string $content, ?int $offset = null, int $mode = 0): bool
+    {
+        $this->checkSupportedModes($mode);
+        $path = $this->fullPath($entry);
+        try {
+            if (FILE_APPEND == $mode) {
+                $handler = @fopen($path, 'ab');
+            } else {
+                $handler = @fopen($path, 'wb');
+            }
+            if (false === $handler) {
+                // @codeCoverageIgnoreStart
+                throw new FilesException($this->getFlLang()->flCannotOpenFile($path));
+            }
+            // @codeCoverageIgnoreEnd
+            if (!is_null($offset)) {
+                if (0 != @fseek($handler, $offset)) {
+                    // @codeCoverageIgnoreStart
+                    throw new FilesException($this->getFlLang()->flCannotSeekFile($path));
+                }
+                // @codeCoverageIgnoreEnd
+            } else {
+                if (!@rewind($handler)) {
+                    // @codeCoverageIgnoreStart
+                    throw new FilesException($this->getFlLang()->flCannotSeekFile($path));
+                }
+                // @codeCoverageIgnoreEnd
+            }
+            if (false === @fwrite($handler, $content)) {
+                // @codeCoverageIgnoreStart
+                throw new FilesException($this->getFlLang()->flCannotWriteFile($path));
+            }
+            // @codeCoverageIgnoreEnd
+            return @fclose($handler);
+        } catch (Throwable $ex) {
+            // @codeCoverageIgnoreStart
+            throw new FilesException($this->getFlLang()->flCannotSaveFile($path), $ex->getCode(), $ex);
+        }
+        // @codeCoverageIgnoreEnd
+    }
+
+    public function saveFileStream(array $entry, $content, int $mode = 0): bool
+    {
+        $this->checkSupportedModes($mode);
+        $path = $this->fullPath($entry);
+        try {
+            if (FILE_APPEND == $mode) { // append to the end
+                $handler = @fopen($path, 'ab');
+            } else { // rewrite all from offset
+                $handler = @fopen($path, 'wb');
+            }
+            if (false === $handler) {
+                // @codeCoverageIgnoreStart
+                throw new FilesException($this->getFlLang()->flCannotOpenFile($path));
+            }
+            // @codeCoverageIgnoreEnd
+            if (false === @stream_copy_to_stream($content, $handler)) {
+                // @codeCoverageIgnoreStart
+                throw new FilesException($this->getFlLang()->flCannotWriteFile($path));
+            }
+            // @codeCoverageIgnoreEnd
+            return @fclose($handler);
+        } catch (Throwable $ex) {
+            // @codeCoverageIgnoreStart
+            throw new FilesException($this->getFlLang()->flCannotSaveFile($path), $ex->getCode(), $ex);
         }
         // @codeCoverageIgnoreEnd
     }
 
     /**
-     * @param string $path
-     * @param resource $content
+     * @param int<0, max> $mode
      * @throws FilesException
      */
-    protected function writeStream(string $path, $content): void
+    protected function checkSupportedModes(int $mode): void
     {
-        if (false === file_put_contents($path, $content)) {
-            // @codeCoverageIgnoreStart
-            throw new FilesException($this->getLang()->flCannotWriteFile($path));
+        if (!in_array($mode, [0, FILE_APPEND])) {
+            throw new FilesException($this->getFlLang()->flBadMode($mode));
         }
     }
 
@@ -130,9 +155,20 @@ class ProcessFile implements IProcessFiles
             return @copy($src, $dst);
             // @codeCoverageIgnoreStart
         } catch (Throwable $ex) {
-            throw new FilesException($this->getLang()->flCannotCopyFile($src, $dst), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotCopyFile($src, $dst), $ex->getCode(), $ex);
         }
         // @codeCoverageIgnoreEnd
+    }
+
+    public function copyFileStream(array $source, array $dest): bool
+    {
+        $stream = $this->readFileStream($source);
+        if (!@rewind($stream)) {
+            // @codeCoverageIgnoreStart
+            throw new FilesException($this->getFlLang()->flCannotSeekFile($this->fullPath($source)));
+        }
+        // @codeCoverageIgnoreEnd
+        return $this->saveFileStream($dest, $stream);
     }
 
     public function moveFile(array $source, array $dest): bool
@@ -143,9 +179,16 @@ class ProcessFile implements IProcessFiles
             return @rename($src, $dst);
             // @codeCoverageIgnoreStart
         } catch (Throwable $ex) {
-            throw new FilesException($this->getLang()->flCannotMoveFile($src, $dst), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotMoveFile($src, $dst), $ex->getCode(), $ex);
         }
         // @codeCoverageIgnoreEnd
+    }
+
+    public function moveFileStream(array $source, array $dest): bool
+    {
+        $r1 = $this->copyFileStream($source, $dest);
+        $r2 = $this->deleteFile($source);
+        return $r1 && $r2;
     }
 
     public function deleteFile(array $entry): bool
@@ -155,7 +198,7 @@ class ProcessFile implements IProcessFiles
             return @unlink($path);
             // @codeCoverageIgnoreStart
         } catch (Throwable $ex) {
-            throw new FilesException($this->getLang()->flCannotRemoveFile($path), $ex->getCode(), $ex);
+            throw new FilesException($this->getFlLang()->flCannotRemoveFile($path), $ex->getCode(), $ex);
         }
         // @codeCoverageIgnoreEnd
     }
@@ -176,6 +219,6 @@ class ProcessFile implements IProcessFiles
      */
     protected function noDirectoryDelimiterSet(): string
     {
-        return $this->getLang()->flNoDirectoryDelimiterSet();
+        return $this->getFlLang()->flNoDirectoryDelimiterSet();
     }
 }
