@@ -5,12 +5,12 @@ namespace VolumeTests;
 
 use CommonTestClass;
 use kalanis\kw_files\FilesException;
-use kalanis\kw_files\Interfaces\IProcessFiles;
+use kalanis\kw_files\Interfaces\IProcessFileStreams;
 use kalanis\kw_files\Processing\Volume\ProcessFile;
 use kalanis\kw_paths\PathsException;
 
 
-class FileTest extends CommonTestClass
+class FileStreamTest extends CommonTestClass
 {
     protected function tearDown(): void
     {
@@ -35,10 +35,7 @@ class FileTest extends CommonTestClass
     public function testRead(): void
     {
         $lib = $this->getLib();
-        $this->assertEquals('qwertzuiopasdfghjklyxcvbnm0123456789', $lib->readFile(['dummy2.txt']));
-        $this->assertEquals('asdfghjklyxcvbnm0123456789', $lib->readFile(['dummy2.txt'], 10));
-        $this->assertEquals('asdfghjkly', $lib->readFile(['dummy2.txt'], 10, 10));
-        $this->assertEquals('qwertzuiop', $lib->readFile(['dummy2.txt'], null, 10));
+        $this->assertEquals('qwertzuiopasdfghjklyxcvbnm0123456789', $this->streamToString($lib->readFileStream(['dummy2.txt'])));
     }
 
     /**
@@ -49,7 +46,7 @@ class FileTest extends CommonTestClass
     {
         $lib = $this->getLib();
         $this->expectException(FilesException::class);
-        $lib->readFile(['unknown']);
+        $lib->readFileStream(['unknown']);
     }
 
     /**
@@ -59,30 +56,7 @@ class FileTest extends CommonTestClass
     public function testSave(): void
     {
         $lib = $this->getLib();
-        $this->assertTrue($lib->saveFile(['extra.txt'], 'qwertzuiopasdfghjklyxcvbnm0123456789'));
-    }
-
-    /**
-     * Insert content
-     * @throws FilesException
-     * @throws PathsException
-     */
-    public function testSave2(): void
-    {
-        $lib = $this->getLib();
-        $this->assertTrue($lib->saveFile(['extra1.txt'], 'qwertzuiopasdfghjklyxcvbnm0123456789'));
-        $this->assertTrue($lib->saveFile(['extra1.txt'], '0123456789', 15));
-    }
-
-    /**
-     * Create file with moved content
-     * @throws FilesException
-     * @throws PathsException
-     */
-    public function testSave3(): void
-    {
-        $lib = $this->getLib();
-        $this->assertTrue($lib->saveFile(['extra2.txt'], 'qwertzuiopasdfgh01234567890123456789', 5));
+        $this->assertTrue($lib->saveFileStream(['extra.txt'], $this->stringToStream('qwertzuiopasdfghjklyxcvbnm0123456789')));
     }
 
     /**
@@ -90,11 +64,11 @@ class FileTest extends CommonTestClass
      * @throws FilesException
      * @throws PathsException
      */
-    public function testSave4(): void
+    public function testSave2(): void
     {
         $lib = $this->getLib();
-        $this->assertTrue($lib->saveFile(['extra3.txt'], 'qwertzuiopasdfgh01234567890123456789', 0, FILE_APPEND));
-        $this->assertTrue($lib->saveFile(['extra3.txt'], 'qwertzuiopasdfgh01234567890123456789', 45, FILE_APPEND));
+        $this->assertTrue($lib->saveFileStream(['extra3.txt'], $this->stringToStream('qwertzuiopasdfgh01234567890123456789'), FILE_APPEND));
+        $this->assertTrue($lib->saveFileStream(['extra3.txt'], $this->stringToStream('qwertzuiopasdfgh01234567890123456789'), FILE_APPEND));
     }
 
     /**
@@ -105,7 +79,7 @@ class FileTest extends CommonTestClass
     {
         $lib = $this->getLib();
         $this->expectException(FilesException::class);
-        $lib->saveFile(['extra.txt'], '0123456789qwertzuiopasdfghjklyxcvbnm', 0, 666);
+        $lib->saveFileStream(['extra.txt'], $this->stringToStream('0123456789qwertzuiopasdfghjklyxcvbnm'), 666);
     }
 
     /**
@@ -115,10 +89,10 @@ class FileTest extends CommonTestClass
     public function testSaveNonWritable(): void
     {
         $lib = $this->getLib();
-        $this->assertTrue($lib->saveFile(['extra.txt'], 'qwertzuiopasdfghjklyxcvbnm0123456789'));
+        $this->assertTrue($lib->saveFileStream(['extra.txt'],  $this->stringToStream('qwertzuiopasdfghjklyxcvbnm0123456789')));
         chmod($this->getTestPath() . DIRECTORY_SEPARATOR . 'extra.txt', 0555);
         $this->expectException(FilesException::class);
-        $lib->saveFile(['extra.txt'], '0123456789qwertzuiopasdfghjklyxcvbnm');
+        $lib->saveFileStream(['extra.txt'],  $this->stringToStream('0123456789qwertzuiopasdfghjklyxcvbnm'));
     }
 
     /**
@@ -128,12 +102,11 @@ class FileTest extends CommonTestClass
     public function testCopyMoveDelete(): void
     {
         $lib = $this->getLib();
-        $this->assertTrue($lib->copyFile(['dummy2.txt'], ['extra1.txt']));
-        $this->assertTrue($lib->moveFile(['extra1.txt'], ['extra2.txt']));
-        $this->assertTrue($lib->deleteFile(['extra2.txt']));
+        $this->assertTrue($lib->copyFileStream(['dummy2.txt'], ['extra1.txt']));
+        $this->assertTrue($lib->moveFileStream(['extra1.txt'], ['extra2.txt']));
     }
 
-    protected function getLib(): IProcessFiles
+    protected function getLib(): IProcessFileStreams
     {
         return new ProcessFile($this->getTestPath());
     }
